#3 Move front end to docs directory
This commit moves the front end to a ./docs/ directory so that not everything's in a big flat pile at the top of the repo.

We kept the *.js files at the top (and out of the docs directory) to make it clear that these aren't being used by the HTML files (they're only used by the process_data.py script that generates those HTML files).