#6 Make Date Changing Speed Zoom-Level Invariant
Since the data takes so long to process, zooming out at a larger level slows down the rate at which the date changes.

If we zoom in, then the date slider changes at a faster pace.

This commit attempts to make this all consistent by slowing down the default speed at which the date changes to one that can be handled when we're zoomed all the way out (i.e. 500ms).